Remove src/logo.svg file
The logo.svg file within the src folder is removed in this commit. This step is necessary as part of the redesign of the application. The traditional logo is no longer suited to the new design direction, resulting in its removal. Further updates will lead to a new, more fitting logo introduced in the future.